refactor: unify color override controls
Extract the full preset, picker, RGB, and automatic-reset control so component foreground/background, wire, and drafting properties share one UI implementation while retaining their separate document protocols. Resolve inherited colors from the active style profile and replace the Violet preset with Gray without changing the remaining palette.

Validation: pnpm gate:preflight -- --base HEAD (static contracts passed; pre-commit Test-Impact check expectedly awaited this commit); pnpm gate:affected -- --base HEAD (298 unit files / 1856 tests and 121 manual-editor browser tests passed); focused drafting and wire color browser checks passed.

Test-Impact: tests-updated
